#20 Add clinic logic and schema
This commit was done by the team together during a group coding session. It completes the last feature for the second milestone. In summary, we've added a new schema called clinic that has a name, location and arrays of people belonging to the clinic. It also adds the logic to be able to create the clinic and to put people inside of it.